feat: merge testsuite elements with the same name (fixes #226, fixes #227)
BREAKING CHANGE: previously we're basically concatenating contents of different input files.
That worked wrong for nested testsuite elements.
In this commit we traverse report XML and merge testsuite elements based on their names.
It is the change in tool behavior, that's why it can be breaking for some use cases.